// src/app/model/employee.model.ts
export interface Employees {
  id: number;
  empFirstName: string;
  empLastName: string;
  empGender: string;
  empDateOfBirth: string;
  empDateOfJoining: string;
  empPhoneNumber: string;
  empEmailId: string;
  empHomeAddrLine1: string;
  empHomeAddrLine2: string;
  empHomeAddrStreet: string;
  empHomeAddrDistrict: string;
  empHomeAddrState: string;
  empHomeAddrCountry: string;
  empHomeAddrPinCode: string;
}